requirements/dev.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.28.83
botocore==1.31.83
certifi==2023.7.22
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.0
Django==4.2.3
django-ace==1.26.0
django-admin-autocomplete-filter==0.7.1
django-admin-extra-buttons==1.5.6
django-jazzmin==2.6.0
django-json-widget==1.1.1
django-coverage-plugin==3.1.0
django-debug-toolbar==4.2.0
django-more-admin-filters==1.7
django-simple-history==3.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==2.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.3
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycurl==7.45.2
python-dateutil==2.8.2
python-decouple==3.8
python-memcached==1.62
pymemcache==4.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.25.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.9
